Introduce basic ownership
For an extern "Rust" `type Foo`, we used to generate a `class Foo` on
the Swift side.

We now generated `class Foo` `class FooRef` and `class FooRefMut`.

A `class Foo` can be used wherever a `Foo` or `&Foo` or `&mut Foo` can be used

A `class FooRefMut` can be used wherever a `&Foo` or `&mut Foo` can be used.

A `class FooRef` can only be used as `&Foo`