Add initial support for async functions (#31)
This commit introduces the ability to expose asynchronous Rust functions
for Swift to call.

Async function support is disabled by default. You can enable it using the
"async" feature, which pulls in `tokio` and `once_cell` as a
dependencies.

```rust
#[swift_bridge::bridge]
mod ffi {
    extern "Rust" {
        async fn some_async_function() -> u32;
    }
}

async fn some_async_function() -> u32 {
    123
}
```

```swift
func testSomeAsyncFunction() async {
    let num = await some_async_function()

    XCTAssertEqual(num, 123)
}
```

---

As a starting point, we're using Swift's built-in async runtime to
drive our async functions.

In the future we'll carefully research which runtime to use (as well as
whether we should allow the user to use a runtime of their choosing).

But, for now, defaulting to Swift's async runtime should work and allow
us to continue iterating on our async support.

---

We do not yet support arguments in async Rust functions.

We do not yet support all of our supported types in async Rust functions.

So far primitives and shared structs are supported as async Rust function
return types.