Re-organize book bridge-module section
This commit re-organizes the bridge module section of the book.

Before this commit we had a section for `extern "Rust"` and another for
`extern "Swift"` modules.

Both of these share a lot of concepts, such as the way that you declare
types and functions, many attributes, and more.

Also, splitting this in this way made it hard to scan the chapter titles
to figure out where to go.

We want people to be able to skip to the chapter that they need for the
task at hand. So, we want to have chapters that are focused around
specific concepts.

As of this commit there are now chapters for functions, opaque types,
and transparent (non-opaque) types. As we continue to flesh out the book
we may split some of these up even further.