Remove rust-analyzer example
`rust-analyzer` takes a lot of time to compile.

The original idea of having a `rust-analyzer` example was
to highlight to users that it was possible to take a big, powerful Rust
crate and easily bridge it.

We could probably accomplish this by just including a short section in
the README and/or book where we talk about example swift-bridge use
cases.