Fix first run of integration tests
This commit makes the `./test-integration.sh` script create empty
versions of the generated bridge files before the Xcodebuild process
starts.

This fixes an issue where if a file is generated during Xcode build, but
was not there before the build process started, you get an error such
as:
`error: Build input file cannot be found: '/path/to/genrated-file'.