Add example of Rust linking in Swift
This commit adds an example of a Rust executable that statically links
to a Swift native library.

Along the way we also make `swift-bridge-build` always generate the
core Swift bridge files, instead of having these files generated
unpredictably as we did before this commit.

---

In the future we can consider abstracting away the swift build command
behind a builder (or some other pattern) that has sane defaults.