Fix integration tests
This commit fixes the `SwiftRustIntegrationTestRunner` project such that
it now builds and runs the full test suite.

Before this commit there was a recently introduced file that was not
actually added to the project, so the project could not compile.

CI was failing for the last few commits so this slipped through the
cracks. This commit should fix CI.