Remove `Array.toUnsafeBufferPointer`
This commit removes the `Array.toUnsafeBufferPointer` usage from the
documentation.
This was a method that we used to have but removed in a previous commit.

Swift pointers have lifetimes that can be shorter than the lifetime of
the object that they are pointing to, so the returned pointer could
unpredictably become a dangling pointer.

The documentation has been updated with an alternative approach to
passing a Swift buffer pointer to Rust.