Document Swift String copying
This commit adds documentation explaining that there is is no zero-copy
way to create a Swift `String`.

We also add documentation explaining that the reason the `RustString`
type exists is to avoid the implicit allocation that would be required
if we automatically converted Rust's `std::string::String` to Swift's
`String`.

We also document some potential new attributes that would allow users to
opt into automatically converted from a `Rust std::string::String` to a
Swift String.